requirements.txt
opt-einsum
keras-preprocessing
gast==0.2.2
google-pasta
astor
keras-applications
wrapt
numpy==1.20.1
scipy==1.4.1
h5py
cython
pytz
opencv-python
tqdm
pyyaml==5.4.1
tensorboard==2.1.0
tabulate
termcolor
tensorflow-estimator
joblib
lick
tensorboardX
easydict
scikit-learn
atari-py
tensorflow
cycler
kiwisolver
python-dateutil
pyparsing
Pillow<7
cloudpickle==1.2.2
future
tensorflow
matplotlib==3.1.3
pyglet==1.3.2
gym==0.15.4
portalocker==1.5.1
yacs==0.1.6
pybullet==3.1.0
